+++ ProgramDataGenerator/src/programdatagenerator/simulationdata/Product.java
@@ -41,7 +41,7 @@ public class Product {
     private int TotalUnitCnt=0;
     
     private List<Lot> RandomLot= new ArrayList<>();
-    private int AvgTestTime=0;
+    private Double AvgTestTime=0.0;
    
 
     @SuppressWarnings("empty-statement")
@@ -63,7 +63,7 @@ public class Product {
                             readValue(this.ProgramVersion,node);
                            break;
                         case "AvgTestTime":
-                           this.AvgTestTime=Integer.valueOf(node.getText());
+                           this.AvgTestTime=Double.valueOf(node.getText());
                            break;
                         case "TestCode":
                             readValue(this.TestCode, node);
